@@ -1,16 +1,16 @@
 name: shelf_proxy
-version: 0.1.0+8-dev
+version: 0.1.0+8-nullsafety.0
 description: A shelf handler for proxying HTTP requests to another server.
 homepage: https://github.com/dart-lang/shelf_proxy
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  http: '>=0.9.0 <0.13.0'
-  path: '>=1.0.0 <2.0.0'
-  pedantic: ^1.0.0
-  shelf: '>=0.5.2 <0.8.0'
+  http: ^0.13.0-nullsafety.0
+  path: ^1.8.0-nullsafety.3
+  pedantic: ^1.10.0-nullsafety.3
+  shelf: ^1.0.0-nullsafety.0
 
 dev_dependencies:
-  test: ^1.3.0
+  test: ^1.16.0-nullsafety.17
